+++ company.el
@@ -300,18 +300,24 @@ This doesn't include the margins and the scroll bar."
                  (const :tag "Two lines" lines)))
 
 (defcustom company-tooltip-align-annotations nil
-  "When non-nil, align annotations to the right tooltip border.
-
-When the value is a number, maintain at least this many spaces between the
-completion text and its annotation."
-  :type '(choice (const :tag "Align to the right" t)
-                 (number :tag "Align to the right with minimum spacing"))
+  "When non-nil, align annotations to the right tooltip border."
+  :type 'boolean
   :package-version '(company . "0.7.1"))
 
 (defcustom company-tooltip-flip-when-above nil
   "Whether to flip the tooltip when it's above the current line."
   :type 'boolean
   :package-version '(company . "0.8.1"))
+
+(defcustom company-tooltip-annotation-padding nil
+  "Non-nil to specify the padding before annotation.
+
+Depending on the value of `company-tooltip-align-annotations', the default
+padding is either 0 or 1 space.  This variable allows to override that
+value to increase the padding.  When annotations are right-aligned, it sets
+the minimum padding, and otherwise just the constant one."
+  :type 'number
+  :package-version '(company "0.9.14"))
 
 (defvar company-safe-backends
   '((company-abbrev . "Abbrev")
@@ -3089,22 +3095,23 @@ If SHOW-VERSION is non-nil, show the version in the echo area."
          (_ (setq value (company-reformat (company--pre-render value))
                   annotation (and annotation (company--pre-render annotation t))))
          (ann-ralign company-tooltip-align-annotations)
+         (ann-padding (or company-tooltip-annotation-padding 0))
          (ann-truncate (< width
                           (+ (length value) (length annotation)
-                             (or ann-ralign 0))))
+                             ann-padding)))
          (ann-start (+ margin
                        (if ann-ralign
                            (if ann-truncate
-                               (+ (length value) ann-ralign)
+                               (+ (length value) ann-padding)
                              (- width (length annotation)))
-                         (length value))))
+                         (+ (length value) ann-padding))))
          (ann-end (min (+ ann-start (length annotation)) (+ margin width)))
          (line (concat left
                        (if (or ann-truncate (not ann-ralign))
                            (company-safe-substring
                             (concat value
-                                    (when (and annotation ann-ralign)
-                                      (company-space-string ann-ralign))
+                                    (when annotation
+                                      (company-space-string ann-padding))
                                     annotation)
                             0 width)
                          (concat
@@ -3332,10 +3339,9 @@ but adjust the expected values appropriately."
 (defun company--create-lines (selection limit)
   (let ((len company-candidates-length)
         (window-width (company--window-width))
-        (company-tooltip-align-annotations
-         (if (eq company-tooltip-align-annotations t)
-             1
-           company-tooltip-align-annotations))
+        (company-tooltip-annotation-padding
+         (or company-tooltip-annotation-padding
+             (if company-tooltip-align-annotations 1 0)))
         left-margins
         left-margin-size
         lines
@@ -3408,9 +3414,9 @@ but adjust the expected values appropriately."
             (setq annotation (string-trim-left annotation))))
         (push (list value annotation left) items)
         (setq width (max (+ (length value)
-                            (if (and annotation company-tooltip-align-annotations)
+                            (if annotation
                                 (+ (length annotation)
-                                   company-tooltip-align-annotations)
+                                   company-tooltip-annotation-padding)
                               (length annotation)))
                          width))))